# [enhancement] Add indents to storage details when profiling a query